do multiplication in floating point to avoid overflow

+++ opensubdiv/osd/spmvDispatcher.h
@@ -211,7 +211,7 @@ public:
     }
 
     virtual inline double SparsityFactor() {
-        return (double) nnz / (double) (m * n);
+        return (double) nnz / ((double) m * (double) n);
     }
 };
